refactor(web): modularize SavedAds page component (REF-504)
- Extract SavedAdCardItems sub-components
- Extract useSavedAdsSort hook
- Reduce SavedAds.tsx from 530 to 384 LOC
- Preserve grid/list view controls, sorting, and unsave functionality
